\name{RcmdrISCSSPlugin-internal}
\title{Internal RcmdrISCSSPlugin objects}
\alias{freqDist.iscss}
\alias{readDTA}
\alias{importSTATA.iscss}
\alias{searchVarLabels.iscss}
\alias{histDiscrete.iscss}
\alias{numSumAll.iscss}
\alias{subsetDataSet.iscss}
\alias{twoWayTable.iscss}
\alias{unalike.iscss}
\alias{ci.iscss}
\alias{confidenceInterval}
\alias{searchVarLabels.iscss}
\alias{searchVarLabels}
\alias{searchVarLabels.data.frame}
\alias{searchVarLabels.tbl_df}
\alias{inspect.iscss}
\alias{inspect}
\alias{inspect.tbl_df}
\alias{inspect.data.frame}
\alias{asFactor.iscss}
\alias{concordant}
\alias{discordant}
\alias{phi}
\alias{lambda}
\alias{ord.gamma}
\alias{ord.somers.d}
\alias{tau.b}
\alias{V}
\alias{simtable}
\alias{simrho}
\alias{makeStats}
\alias{plotStdRes}
\alias{twoWayTable.iscss}
\alias{pwCorr.iscss}
\alias{pwCorrMat}
\alias{print.pwc}
\alias{pmc}
\alias{plotModelCorrs.iscss}
\description{Internal RcmdrDemosPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
