\name{RcmdrPlugin.ISCSS-package}
\alias{RcmdrPlugin.ISCSS-package}
\alias{RcmdrPlugin.ISCSS}
\docType{package}
\title{
Install the An Introduction to Statistics for Canadian Social Scientists Rcmdr Plug-In
}
\description{
Provides an Rcmdr "plug-in" to support Haan and Godley's An Introduction to Statistics for Canadian Social Scientists.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.ISCSS\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2017-09-25\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Dave Armstrong <dave@quantoid.net>

Maintainer: Dave Armstrong <dave@quantoid.net>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

\examples{
if (interactive()){ # load the Rcmdr and the plug-in
    library(RcmdrPlugin.ISCSS)
    }
}
