\name{RcmdrISCSSPlugin-internal}
\title{Internal RcmdrISCSSPlugin objects}
\alias{freqDist.iscss}
\alias{histDiscrete.iscss}
\alias{numSumAll.iscss}
\alias{subsetDataSet.iscss}
\alias{twoWayTable.iscss}
\alias{unalike.iscss}
\alias{ci.iscss}
\alias{confidenceInterval}
\description{Internal RcmdrDemosPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
